feature: Add rudimentary, half-baked attempt at rendering basic WL graphics descriptions using Skia
This is intended as an intermediate step to moving more of the graphics rendering logic from Skia
to native WL graphics.

At the moment, WL graphics have one subtle but significant drawback from using a graphics
rendering library like Skia.

WL graphics are vector based: coordinates in graphics primitives
like Point[..] or Line[..] or Rectangle[..] are unitless. This has the benefit that the results
graphics can be drawn at any physical size any remain perfectly crisp.

However, the drawback is that the only way to specify the size of text within a graphics
is in printers points. This means that text in a WL graphic will have the same physical
on-screen size no matter how large of small the resulting graphic is rendered. This makes it
extremely difficult to take the width and height of a rendering string of text into account
when trying to compute an automated layout for the elements in the graphic.

For graphics where each string of text is relatively small, this is not always significant.
However, it is significant in graphics with larger amounts of text, or where some elements
have lots of text and others have a short amount of text.