feature: Port layout and rendering code from Rust to WL
This uses the ad hoc Rust code for rendering WL-like graphics
specifications added in a previous commit to do diagram
rendering. The logic for turning a Diagram[..] into a
PlacedDiagram[..] (LayoutDiagram), and then finally into a
sequence of graphics commands (RenderPlacedDiagramToGraphics)
now resides in WL instead of Rust.

This opens the door to experimenting with fancier layout
algorithms that will be easier to prototype in WL than in
Rust.

Illustrative Example:

    basic = Diagram[
        {DiaBox["Hello"], DiaBox["World"]},
        {DiaArrow["Hello" -> "World", "label"]}
    ];

    placed = LayoutDiagram[basic];

    graphics = RenderPlacedDiagramToGraphics[placed];

    DiagramGraphicsImage[Graphics[N@Flatten@graphics]]

* Add SizedText graphics primitive
* Add Primitive::SizedText(..)
* Add PlacedDiagram[..], LayoutDiagram[..], RenderPlacedDiagramToGraphics[..], and RenderedTextSize[..] WL symbols for creating, laying out, and rendering diagrams.
* Add new DiagramMaker`Errors` module with basic exception-based error handling functions