feature: Add a new "RowsLayout" diagram layout method
This diagram layout method is roughly a combination of the existing "RowLayout"
and "GraphLayout" algorithms. WL GraphEmbedding is used to determine a basic
layout for the graph elements (but which does not take into account the size of
each vertex in any way) and then a modifiation of the simple offset logic in
the "RowLayout" algorithm is used to position the boxes with appropriate spacing.

This is the first diagram layout that produces not trivially poor results.