polish: Include DiaArrow labels as tooltips in the diagram, for now
This has the significant downside that these tooltips will not be visible
if the diagram is exported as a static image. They also aren't particularly
discoverable at the moment.

However, implementing support for placing arrow labels during layout will
require a bit of work, so this a quick and dirty way to get these into the
diagram for the time being.