feature: Add Axes option to Diagram
This option is forwarded to the Graphics[..] used to render the diagram.

Setting this option to True is useful for debugging the diagram layout.

If $DebugDiagramLayout is set to True and the Axes option is set to
its default value of Automatic, then `Axes -> True` will be set on
the diagram Graphics.