feature: Add support non-text 'content' in DiaBox's
DiaBox previously supported a 2nd argument that could be an atritrary
string that would be displayed as the content of the DiaBox.

This commit begins to add support for more arbitrary content inside
of DiaBox's. The first new form to be supported is Column, which can
be used for arranging pieces of text to be layed out vertically within
the rendered DiaBox.

Illustrative Example:

	Diagram[
		{DiaBox["Greeting", Column[{"Hello", "World"}]]},
		{}
	]

	will display as a diagram with a single box, containing the
	words "Hello" and "World" on separate lines.

Changes:

* Add new DiaPaclet[..] box type, for representing paclets.
  - The paclet name and version number will be shown in different
    Column items.
* Replace DiagramElementText with DiagramElementContent
* Add LayoutBoxContent, which will eventually replace MakeBoxRectangles.

  - MakeBoxRectangles only understood a string label for a box.
    LayoutBoxContent is intended to handle more complex input, composed
    of arranged text, graphics, and eventually nested diagram elements.

* Add Bounded[..] symbolic form
* Add AddUnmatchedArgumentsHandler[..] utility
* Add RectangleBoundingBox[..] utility
* Add RectangleCenter[..] utility