feature: Add NotebookImportCell utility
This isn't directly used by Diagrams (yet), but its useful for experimenting
with a good UX for "program diagrams", in conjunction with a NotebookWebsiteTools
notebook.

The intended use case is something like a HighlightSyntax cell containing
source code, and then a ProgramDiagram[PreviousCell[], ...]-like call in the
next cell.

This construct avoids the need to write a program as a string literal, as in
something like ProgramDiagram["fn foo() { ... }", ...] which is not a fun way
to imagine writing diagrammed code.