feature: Introduce new Diagram[<| ... |>] object form
This form will typically contain a "Graphics" and "Regions" fields.

Using a dedicated symbolic form will make composition of diagram
functions easier, as well as make it easier for users to perform
operations on Diagrams.

For example, I'm hoping to follow this up with a DiagramAnnotate[_Diagram, _]
function for adding structured annotations (e.g. arrows, brackets, callouts,
etc.) to diagrams in a consistent way.