cleanup: Add new Diagrams`Kinds` module + move BlockStack, BinaryLayout, and Program diagrams under it
The organization of this paclet has suffered a bit since I picked work
up on this project again earlier this year. There are now effectively
two entirely separate representations of Diagrams (the 'boxes and
arrows' model I originally used when I started ~2.5 years ago; and the
*Diagram[..] functions that I've started adding recently).

This doesn't aim to clean up that issue, but it does aim to consolidate
the logic for the latter kind of diagrams in the Kinds/ directory.