feature: Support {"Custom", _List} output spec + LabelingFunction on FileSystemTreeDiagram
This makes it easy to customize the output of a FileSystemTreeDiagram,
e.g. to change the 4-char line "cells", or add a label to a node.

This is useful for "meta" diagrams that describe how FileSystemTreeDiagram
works, like I wrote in connorgray.com/ephemera/worklog describing the
file tree rendering logic used in this function.