feat: Add new "blob" testing framework + initial tests
This adds a new directory with the structure:

    BlobTestsData
    |-- Test1
    |   |-- Input.wl
    |   \-- ExpectedOutput.wl
    |-- Test2
    |   |-- Input.wl
    |   \-- ExpectedOutput.wl
    \-- ...

and a new TestingUtils`RunBlobTests function.

These tests are designed to make it easy to store the entire
raw "blob" form of Diagram expression in a git-friendly way.

This is a much deeper kind of test than the existing diagrams
tests, which only check specified properties of the diagram,
but not its underlying Graphics expression. (Hand written
tests against a Graphics expression are likely to be fragile
and laborious to update.)

Now, if the Graphics "blob" form for a given test `Input.wl`
file changes, the human can review the rendered _visual_ form
of the expression to see if the change is correct.

This keeps the tests meaningful and easy to understand, while
also being robust against unintended changes slipping through
without notice.

Notably, this change already caught several minor bugs in a
recent refactoring.